// Features/payrollConfiguration/Types/IPayrollConfiguration.ts
export interface IPayrollConfiguration {
    idPayrollConfiguration: number;
    topAFP: number;
    topARL: number;
    topARS: number;
    percentageARS?: number;
    percentageAFP?: number;
    arsToExpensePercentage?: number;
    afpToExpensePercentage?: number;
    arlToPercentage?: number;
    percentageARL?: number;
    sindicatePercentageSpecial?: number;
    minimumSalary?: number;

    dataAreaId?: string;
    journalPayrollName?: string;
    accountInvoiceTrampoline?: string;
    payInvoiceJournalName?: string;
    payCheckForPayrollNumber?: string;
    payCheckForFiredNumber?: string;
    paymentMethodForPayroll?: string;
    paymentMethodForFired?: string;
    offsetAccountDisplayValueForPayroll?: string;
    offsetAccountDisplayValueForFired?: string;
    checkJournalNameForPayroll?: string;
    checkJournalNameForFired?: string;
    dynamicURL?: string;
    dynamicClientId?: string;
    dynamicClientSecretId?: string;
    loginURL?: string;
    payBankJournalBatchName?: string;
    ciEmployeeStoreUrl?: string;
    ciAttendancePunchUrl?: string;
    start: Date;
    end: Date;

    sindicatePercentage?: number;
    infotepPercentage?: number;
    infotepSpecial?: number;
    bonificationBreak?: string[];
    seniorityIncentiveBreak?: string[];
    isrBreak: string[];
    isr: number;
    salary: number;
    afp: number;
    ars: number;
    arl: number;
    infotep: number;
    bonification: number;
    christmasSalary: number;
    unemployment: number;
    notice: number;
    isrInFavor: number;
    outstandingDebt: number;
    vacation: number;
    previousPeriodDebt: number;
    vacationIncentive: number;
    salaryToPayEmployee: number;
    salaryToPayCompany: number;
    afpCompanyDebit: number;
    arsCompanyDebit: number;
    arlCompanyDebit: number;
    bonificationCompany: number;
    christmasSalaryCompany: number;
    infotepCompany: number;
    minTaxByEmployee?: boolean;
    afpCompanyCredit: number;
    arsCompanyCredit: number;
    arlCompanyCredit: number;
    childSupport?: number;
}